api/src/app/requirements.txt
flask-restful
flask-sqlalchemy
flask-migrate
flask-mail
flask-cors
mysqlclient
uwsgi
pytest
pytest-cov
requests
schedule
dropbox
